backend/requirements.txt
# Core FastAPI and async support
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Database and caching
sqlalchemy==2.0.23
asyncpg==0.29.0
redis==5.0.1
alembic==1.13.0

# AI and vector search
openai==1.3.6
pydantic-ai==0.0.14
chromadb==0.4.18
markdown-it-py==3.0.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pylint==4.0.4

# Environment and configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# WebSocket support
websockets==12.0